Create useCancellable comp. + use it in Home
